move st.video and st.audio unit tests into their own files
